Set root import path on instantiation
